+++ src/test/java/com/epam/reportportal/junit5/CodeReferenceTest.java
@@ -48,14 +48,18 @@ public class CodeReferenceTest {
 
 		Launch launch = CodeReferenceTestExtension.LAUNCH;
 
-		verify(launch, times(1)).startTestItem(any()); // Start parent Suite
+		ArgumentCaptor<StartTestItemRQ> suiteCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
+		verify(launch, times(1)).startTestItem(suiteCaptor.capture()); // Start parent Suite
 
-		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
-		verify(launch, times(1)).startTestItem(notNull(), captor.capture()); // Start a test
+		ArgumentCaptor<StartTestItemRQ> testCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
+		verify(launch, times(1)).startTestItem(notNull(), testCaptor.capture()); // Start a test
 
-		List<StartTestItemRQ> rqValues = captor.getAllValues();
 		String className = SingleTest.class.getCanonicalName();
-		assertThat(rqValues.get(0).getCodeRef(), equalTo(className + ".singleTest"));
+		StartTestItemRQ suiteRq = suiteCaptor.getValue();
+		assertThat(suiteRq.getCodeRef(), equalTo(className));
+
+		StartTestItemRQ testRq = testCaptor.getValue();
+		assertThat(testRq.getCodeRef(), equalTo(className + ".singleTest"));
 	}
 
 	@Test
@@ -64,13 +68,18 @@ public class CodeReferenceTest {
 
 		Launch launch = CodeReferenceTestExtension.LAUNCH;
 
-		verify(launch, times(1)).startTestItem(any()); // Start parent Suite
+		ArgumentCaptor<StartTestItemRQ> suiteCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
+		verify(launch, times(1)).startTestItem(suiteCaptor.capture()); // Start parent Suite
 
-		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
-		verify(launch, times(2)).startTestItem(notNull(), captor.capture()); // Start a test class and a test
+		ArgumentCaptor<StartTestItemRQ> testCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
+		verify(launch, times(2)).startTestItem(notNull(), testCaptor.capture()); // Start a test class and a test
 
-		List<StartTestItemRQ> rqValues = captor.getAllValues();
-		String testName = SingleDynamicTest.class.getCanonicalName() + ".testForTestFactory";
+		String className = SingleDynamicTest.class.getCanonicalName();
+		StartTestItemRQ suiteRq = suiteCaptor.getValue();
+		assertThat(suiteRq.getCodeRef(), equalTo(className));
+
+		List<StartTestItemRQ> rqValues = testCaptor.getAllValues();
+		String testName = className + ".testForTestFactory";
 		assertThat(rqValues.get(0).getCodeRef(), equalTo(testName));
 		assertThat(rqValues.get(1).getCodeRef(), equalTo(testName + "$" + SingleDynamicTest.TEST_CASE_DISPLAY_NAME));
 	}